lua.cmake: fix add_lua_test CMAKE_CFG_INTDIR expansion in module paths.
In makefile-based generators, the generator expression `$<CONFIGURATION>`
(used in `add_test`) will be replaced by the value of
`CMAKE_BUILD_TYPE`.  The variable `CMAKE_CFG_INTDIR` is `.`.

In multiple-configuration based generators, like for msvc,
`CMAKE_BUILD_TYPE` is ignored, and `$<CONFIGURATION>` will be
replaced by the (mandatory) `-C` argument passed to ctest.
Binary files are written to `${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}`,
where `${CMAKE_CFG_INTDIR}` is an expression whose value is (perhaps)
not known to cmake but is expanded by the native build tool.
In the case of msvc, this directory seems equivalent to
`${CMAKE_CURRENT_BINARY_DIR}/$<CONFIGURATION>`, but I'm not sure
that this is the case in general and can be relied upon.

Previously, the Lua wrapper generated by `lua.cmake:add_lua_test` was
substituting `CMAKE_CFG_INTDIR` in module paths with `$<CONFIGURATION>`,
which is wrong for (at least) makefile-based generators.
It will cause tests to fail since LuaDist/Bootstrap/CMakeLists.txt
defaults ` CMAKE_BUILD_TYPE` to `"MinSizeRel"`). We might allow
`lua.cmake` to expand `${CMAKE_CFG_INTDIR}` in modules paths to a
native build variable (e.g. `$(OutDir)` for vs9 or `.` for
makefile generators) and then the Lua wrapper will expand the native
build variable to an actual path--e.g. replacing `$(OutDir)`
(but not `.`) with the value of `$<CONFIGURATION>` (which, as I said,
seems to work in practice), but on second thought, note what the cmake
manual says of variables like `$(OutDir)`:

> "Since these values are evaluated by the native build system,
> this variable is suitable only for use in command lines that
> will be evaluated at build time."

That we understand to mean that variables like `$(OutDir)` should be
passed on the command line so that the native build tool will
properly expand them (not the Lua wrapper).
This implies that the `lua.cmake:add_lua_test add_test` command should
be written like this:

```
add_test ( NAME ${testname} COMMAND ${_pre} "${LUA}"
    "${TESTWRAPPER}" "${CMAKE_CFG_INTDIR}" "${_ARG_DEFAULT_ARGS}" )
```